+++ example/lib/src/view/home/home_page.dart
@@ -9,7 +9,7 @@ import 'package:example/src/view.dart';
 /// The Home page
 class HomePage extends StatefulWidget {
   /// With a constant constructor, instantiated once.
-  const HomePage({super.key, this.title = 'Flutter InheritedWidget Demo'});
+  const HomePage({super.key, this.title = 'Hello! example'});
 
   /// Fields in a StatefulWidget should always be "final".
   final String title;
@@ -35,26 +35,13 @@ class _HomePageState extends StateX<HomePage> {
     /// Look inside the parent function and see it calls
     /// all it's Controllers if any.
     super.initState();
-
-    var con = controller;
-
-    con = controllerById(con?.identifier);
-
-    /// Retrieve the 'app level' State object
-    appState = rootState!;
-
-    /// You're able to retrieve the Controller(s) from other State objects.
-    con = appState.controller;
-
-    con = appState.controllerById(con?.identifier);
   }
-
-  late AppStateX appState;
 
   /// The State class' interface
   @override
   Widget build(BuildContext context) {
-    final textStyle = Theme.of(context).textTheme.headlineMedium;
+    //
+    final textTheme = Theme.of(context).textTheme;
     return Scaffold(
       appBar: AppBar(
         title: Text(widget.title),
@@ -76,17 +63,20 @@ class _HomePageState extends StateX<HomePage> {
                   key: const Key('greetings'),
                   style: TextStyle(
                     color: Colors.red,
-                    fontSize: textStyle?.fontSize,
+                    fontSize: textTheme.headlineMedium?.fontSize,
                   ),
                 ),
               ),
             ),
             Text(
               'You have pushed the button this many times:',
-              style: Theme.of(context).textTheme.bodyLarge,
+              style: textTheme.bodyLarge,
             ),
-            // A Text widget to display the counter is in here.
-            CounterWidget(controller: con),
+            // The count is here.
+            Text(
+              '${con.data}',
+              style: textTheme.headlineMedium,
+            ),
           ],
         ),
       ),
@@ -97,42 +87,6 @@ class _HomePageState extends StateX<HomePage> {
         onPressed: con.incrementCounter,
         child: const Icon(Icons.add),
       ),
-    );
-  }
-}
-
-/// Demonstrating the InheritedWidget's ability to spontaneously rebuild
-/// its dependent widgets.
-class CounterWidget extends StatefulWidget {
-  /// Pass along the State Object Controller to make this widget
-  /// dependent on the App's InheritedWidget.
-  const CounterWidget({super.key, this.controller});
-
-  /// Making this widget dependent will cause the build() function below
-  /// to run again if and when the App's InheritedWidget calls its notifyClients() function.
-  final Controller? controller;
-
-  @override
-  State<StatefulWidget> createState() => _CounterState();
-}
-
-class _CounterState extends StateX<CounterWidget> {
-  @override
-  void initState() {
-    super.initState();
-    con = widget.controller;
-  }
-
-  Controller? con;
-
-  @override
-  Widget build(BuildContext context) {
-    // Set a dependency to this widget
-//    con?.dependOnInheritedWidget(context);
-    rootState?.dependOnInheritedWidget(context);
-    return Text(
-      '${con?.count ?? 0}',
-      style: Theme.of(context).textTheme.headlineMedium,
     );
   }
 }